@@ -1,31 +1,26 @@
-﻿using JsonSubTypes;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Converters;
-using Newtonsoft.Json.Serialization;
+﻿using System.Text.Json;
+using System.Text.Json.Serialization;
 
 namespace TDAmeritradeAPI
 {
     public class Order : ICloneable
     {
+        private readonly JsonSerializerOptions _jsonSerializerOptions;
         public class Enums
         {
             // This enum is only for the OrderGenerator class to know if the symbol in question is from an equity, an option, or something else
             public enum SecurityType { Equity, Option, Other }
 
-            [JsonConverter(typeof(StringEnumConverter))]
             public enum Session { NORMAL, AM, PM, SEAMLESS }
 
-            [JsonConverter(typeof(StringEnumConverter))]
             public enum Duration { DAY, GOOD_TILL_CANCEL, FILL_OR_KILL }
 
-            [JsonConverter(typeof(StringEnumConverter))]
             public enum OrderType
             {
                 MARKET, LIMIT, STOP, STOP_LIMIT, TRAILING_STOP, MARKET_ON_CLOSE,
                 EXERCISE, TRAILING_STOP_LIMIT, NET_DEBIT, NET_CREDIT, NET_ZERO
             }
 
-            [JsonConverter(typeof(StringEnumConverter))]
             public enum ComplexOrderStrategyType
             {
                 NONE, COVERED, VERTICAL, BACK_RATIO, CALENDAR, DIAGONAL,
@@ -34,44 +29,32 @@ namespace TDAmeritradeAPI
                 UNBALANCED_VERTICAL_ROLL, CUSTOM
             }
 
-            [JsonConverter(typeof(StringEnumConverter))]
             public enum RequestedDestination { INET, ECN_ARCA, CBOE, AMEX, PHLX, ISE, BOX, NYSE, NASDAQ, BATS, C2, AUTO }
 
-            [JsonConverter(typeof(StringEnumConverter))]
             public enum PriceLinkBasis { MANUAL, BASE, TRIGGER, LAST, BID, ASK, ASK_BID, MARK, AVERAGE }
 
-            [JsonConverter(typeof(StringEnumConverter))]
             public enum PriceLinkType { VALUE, PERCENT, TICK }
 
-            [JsonConverter(typeof(StringEnumConverter))]
             public enum StopType { STANDARD, BID, ASK, LAST, MARK }
 
-            [JsonConverter(typeof(StringEnumConverter))]
             public enum TaxLotMethod { FIFO, LIFO, HIGH_COST, LOW_COST, AVERAGE_COST, SPECIFIC_LOT }
 
-            [JsonConverter(typeof(StringEnumConverter))]
             public enum OrderLegType { EQUITY, OPTION, INDEX, MUTUAL_FUND, CASH_EQUIVALENT, FIXED_INCOME, CURRENCY }
 
-            [JsonConverter(typeof(StringEnumConverter))]
             public enum Instruction
             {
                 BUY, SELL, BUY_TO_COVER, SELL_SHORT, BUY_TO_OPEN, BUY_TO_CLOSE, SELL_TO_OPEN,
                 SELL_TO_CLOSE, EXCHANGE
             }
 
-            [JsonConverter(typeof(StringEnumConverter))]
             public enum PositionEffect { OPENING, CLOSING, AUTOMATIC }
 
-            [JsonConverter(typeof(StringEnumConverter))]
             public enum QuantityType { ALL_SHARES, DOLLARS, SHARES }
 
-            [JsonConverter(typeof(StringEnumConverter))]
             public enum SpecialInstruction { ALL_OR_NONE, DO_NOT_REDUCE, ALL_OR_NONE_DO_NOT_REDUCE }
 
-            [JsonConverter(typeof(StringEnumConverter))]
             public enum OrderStrategyType { SINGLE, OCO, TRIGGER }
 
-            [JsonConverter(typeof(StringEnumConverter))]
             public enum Status
             {
                 AWAITING_PARENT_ORDER, AWAITING_CONDITION, AWAITING_MANUAL_REVIEW, ACCEPTED,
@@ -110,9 +93,6 @@ namespace TDAmeritradeAPI
             public double? Quantity { get; set; }
             public Enums.QuantityType? QuantityType { get; set; }
         }
-
-        // Private variables
-        private static readonly JsonSerializerSettings _serializerSettings = new();
 
         // Properties
         public Enums.Session? Session { get; set; }
@@ -153,14 +133,22 @@ namespace TDAmeritradeAPI
         public string? StatusDescription { get; set; }
         public Order()
         {
-            _serializerSettings.ContractResolver = new CamelCasePropertyNamesContractResolver();
-            _serializerSettings.NullValueHandling = NullValueHandling.Ignore;
+            _jsonSerializerOptions = new JsonSerializerOptions()
+            {
+                PropertyNamingPolicy = JsonNamingPolicy.CamelCase,
+                WriteIndented = true,
+                Converters =
+                {
+                    new JsonStringEnumConverter()
+                }
+            };
         }
-        public string AsJson()
+
+        public string ToJson()
         {
-            string jsonString = JsonConvert.SerializeObject(this, _serializerSettings);
-            return jsonString;
+            return JsonSerializer.Serialize(this, _jsonSerializerOptions);
         }
+
         public object Clone()
         {
             Order clone = (Order)MemberwiseClone();
@@ -172,15 +160,16 @@ namespace TDAmeritradeAPI
         }
         internal static Order GetChildOrderStrategy(Order order)
         {
-            Order newOrder = new() {
+            Order newOrder = new()
+            {
                 OrderType = order.OrderType,
                 Session = order.Session,
                 Price = order.Price,
                 Duration = order.Duration,
                 OrderStrategyType = order.OrderStrategyType,
                 OrderLegCollection = order.OrderLegCollection is not null ? new List<OrderLeg>(order.OrderLegCollection) : null
-        };
-            
+            };
+
             return newOrder;
         }
     }
@@ -190,10 +179,8 @@ namespace TDAmeritradeAPI
         //Enums
         public class Enums
         {
-            [JsonConverter(typeof(StringEnumConverter))]
             public enum ActivityType { EXECUTION, ORDER_ACTION }
 
-            [JsonConverter(typeof(StringEnumConverter))]
             public enum ExecutionType { FILL }
         }
         // Structs
@@ -214,18 +201,11 @@ namespace TDAmeritradeAPI
         public double? OrderRemainingQuantity { get; set; }
         public List<Structs.ExecutionLeg>? ExecutionLegs { get; set; }
     }
-
-    [JsonConverter(typeof(JsonSubtypes), "AssetType")]
-    [JsonSubtypes.KnownSubType(typeof(Equity), AssetTypes.EQUITY)]
-    [JsonSubtypes.KnownSubType(typeof(FixedIncome), AssetTypes.FIXED_INCOME)]
-    [JsonSubtypes.KnownSubType(typeof(MutualFund), AssetTypes.MUTUAL_FUND)]
-    [JsonSubtypes.KnownSubType(typeof(CashEquivalent), AssetTypes.CASH_EQUIVALENT)]
-    [JsonSubtypes.KnownSubType(typeof(Option), AssetTypes.OPTION)]
     public class Instrument
     {
         // Enums
-        [JsonConverter(typeof(StringEnumConverter))]
         public enum AssetTypes { EQUITY, OPTION, INDEX, MUTUAL_FUND, CASH_EQUIVALENT, FIXED_INCOME, CURRENCY }
+
         // Properties
         public Instrument() { }
         public Instrument(string? symbol, AssetTypes? assetType)
@@ -251,10 +231,8 @@ namespace TDAmeritradeAPI
 
     public class MutualFund : Instrument
     {
-        // Enums
         public class Enums
         {
-            [JsonConverter(typeof(StringEnumConverter))]
             public enum Type { NOT_APPLICABLE, OPEN_END_NON_TAXABLE, OPEN_END_TAXABLE, NO_LOAD_NON_TAXABLE, NO_LOAD_TAXABLE }
         }
         // Properties
@@ -263,10 +241,8 @@ namespace TDAmeritradeAPI
 
     public class CashEquivalent : Instrument
     {
-        // Enums
         public class Enums
         {
-            [JsonConverter(typeof(StringEnumConverter))]
             public enum Type { SAVINGS, MONEY_MARKET_FUND }
         }
         // Properties
@@ -278,15 +254,11 @@ namespace TDAmeritradeAPI
         // Enums
         public class Enums
         {
-            [JsonConverter(typeof(StringEnumConverter))]
             public enum Type { VANILLA, BINARY, BARRIER }
-
-            [JsonConverter(typeof(StringEnumConverter))]
             public enum PutCall { PUT, CALL }
-
-            [JsonConverter(typeof(StringEnumConverter))]
             public enum CurrencyType { USD, CAD, EUR, JPY }
         }
+
         // Structs
         public class Structs
         {
@@ -304,5 +276,6 @@ namespace TDAmeritradeAPI
         public string? UnderlyingSymbol { get; set; }
         public int? OptionMultiplier { get; set; }
         public List<Structs.OptionDeliverable>? OptionDeliverables { get; set; }
+        public Option() { }
     }
 }
